+++ src/sora/sora_client.h
@@ -11,9 +11,7 @@
 #include <boost/asio/io_context.hpp>
 #include <boost/asio/ip/tcp.hpp>
 #include <boost/asio/ssl/context.hpp>
-
-// nlohmann/json
-#include <nlohmann/json.hpp>
+#include <boost/json.hpp>
 
 #include "rtc/rtc_manager.h"
 #include "rtc/rtc_message_sender.h"
@@ -32,8 +30,8 @@ struct SoraClientConfig {
   std::string audio_codec_type = "";
   int video_bit_rate = 0;
   int audio_bit_rate = 0;
-  nlohmann::json metadata;
-  nlohmann::json signaling_notify_metadata;
+  boost::json::value metadata;
+  boost::json::value signaling_notify_metadata;
   std::string role = "sendonly";
   bool multistream = false;
   bool spotlight = false;
@@ -76,7 +74,7 @@ class SoraClient : public std::enable_shared_from_this<SoraClient>,
   void DoSendPong();
   void DoSendPong(
       const rtc::scoped_refptr<const webrtc::RTCStatsReport>& report);
-  void CreatePeerFromConfig(nlohmann::json jconfig);
+  void CreatePeerFromConfig(boost::json::value jconfig);
 
  private:
   void OnConnect(boost::system::error_code ec);